feat: Move all Sumcheck engine related items to the sumcheck module
- Introduced a new module `src/spartan/sumcheck/engine.rs`,
- Changed the visibility status of the `sumcheck` module, making it visible within the `spartan` module.
- Refactored the file structure as `src/spartan/sumcheck.rs` was renamed to `src/spartan/sumcheck/mod.rs` and a new submodule `engine` was added within the `spartan` crate.
- Refactored the import structures in `ppsnark` and `sumcheck::engine`, moving several items to `sumcheck::engine`.
