+++ src/spartan/mod.rs
@@ -15,7 +15,7 @@ pub(crate) mod math;
 pub mod polys;
 pub mod ppsnark;
 pub mod snark;
-mod sumcheck;
+pub(in crate::spartan) mod sumcheck;
 
 use crate::{
   r1cs::{R1CSShape, SparseMatrix},
